apiRouter now always checks for authorization, and sets req.user if the token is good
